Move files out of testapp folder
