+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/resources/view/ResourceListFragment.kt
@@ -1,10 +1,10 @@
-package org.wycliffeassociates.otter.jvm.resourcestestapp.view
+package org.wycliffeassociates.otter.jvm.app.ui.resources.view
 
 import org.wycliffeassociates.otter.jvm.app.ui.mainscreen.view.MainScreenStyles
 import org.wycliffeassociates.otter.jvm.app.widgets.workbookheader.workbookheader
 import org.wycliffeassociates.otter.jvm.app.widgets.resourcecard.styles.ResourceListStyles
 import org.wycliffeassociates.otter.jvm.app.widgets.resourcecard.view.ResourceListView
-import org.wycliffeassociates.otter.jvm.resourcestestapp.viewmodel.ResourcesViewModel
+import org.wycliffeassociates.otter.jvm.app.ui.resources.viewmodel.ResourcesViewModel
 import tornadofx.*
 
 class ResourceListFragment : Fragment() {
